+++ data-raw/gsutil_path_parsing.R
@@ -31,7 +31,7 @@ write_with_path_name = function(actual_data_object = NULL,
   all_file_header = "human-precovid-sed-adu" #this is the base structure for all files within the phase.
   tissue_code = .match_ome_tissue_code(desired_ome = ome, input_tissue = tissue)
   file_name = paste(all_file_header, tissue_code, ome, data_category, data_details, sep = "_")
-  file_name = paste0(local_path, file_name, "_v", version, file_type)
+  file_name = paste0(local_path,"/", file_name, "_v", version, file_type)
   if(return_name_only) {
     return(file_name)
   }else{
@@ -105,6 +105,20 @@ write_with_path_name = function(actual_data_object = NULL,
   if(nrow(specific_output) == 0) stop("No ome matches the desired ome/tissue combo")
   tissue_name_output = specific_output[["tissue_code"]]
   return(tissue_name_output)
+}
+
+
+.validate_staging_folder = function(){
+  current_staging_folder = "gs://pre-cawg/staging_20260428"
+  system(paste("gsutil cp -R gs://motrpac-data-hub/analysis/human-precovid-sed-adu/v1.3/epigenomics", current_staging_folder))
+  system(paste("gsutil cp -R gs://motrpac-data-hub/analysis/human-precovid-sed-adu/v1.3/metabolomics-targeted", current_staging_folder))
+  system(paste("gsutil cp -R gs://motrpac-data-hub/analysis/human-precovid-sed-adu/v1.3/metabolomics-untargeted", current_staging_folder))
+  system(paste("gsutil cp -R gs://motrpac-data-hub/analysis/human-precovid-sed-adu/v1.3/proteomics", current_staging_folder))
+  system(paste("gsutil cp -R gs://motrpac-data-hub/analysis/human-precovid-sed-adu/v1.3/resources", current_staging_folder))
+  system(paste("gsutil cp -R gs://motrpac-data-hub/analysis/human-precovid-sed-adu/v1.3/transcriptomics", current_staging_folder))
+  system(paste("gsutil cp -R gs://motrpac-data-hub/analysis/human-precovid-sed-adu/v1.3/clinical_chemistry", current_staging_folder))
+
+
 }